+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Added
 - New models: Qwen3, Gemma4, DeepSeek R1 Distill Qwen
 - Support for Python 3.12 and 3.13
+- Basic model testing (evaluation) functionality.
 
 ### Changed
 - Replace the Llama specific generator with a generic code + config based